mcp-servers/reference-reviewer/pyproject.toml
[project]
name = "reference-reviewer"
version = "0.1.0"
description = "MCP server that generates interactive HTML reference viewers for markdown manuscripts"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pymupdf4llm>=0.0.17",
    "pymupdf>=1.25.0",
    "biopython>=1.84",
    "habanero>=1.2.6",
]
